+++ scripts/dynamic_box.py
@@ -17,9 +17,7 @@ def deform_mesh(frame):
         vertex[1] += (random() - 0.5) * 0.1
         vertex[2] += (random() - 0.5) * 0.1
         mesh.vertex_attributes(v, "xyz", vertex)
-    obj.init() # TODO: this should called, fix later
     obj.update()
-    print(frame)
 
 
 viewer.show()