+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Fixed `action` bug.
+* Updated `compas_viewer.scene.ViewerScene` to support save/load and switching.
+* Changed `is_visible` parameters to `show` in `compas_viewer.scene.ViewerSceneObject` classes.
+* Changed `GridObject` to draw direcly in `Render`
 
 ### Removed